put all flags impls in a single module
