+++ lib/segment/benches/dynamic_mmap_flags.rs
@@ -5,8 +5,8 @@ use std::sync::atomic::AtomicBool;
 use criterion::{Criterion, criterion_group, criterion_main};
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
+use segment::common::flags::dynamic_mmap_flags::DynamicMmapFlags;
 use segment::common::operation_error::check_process_stopped;
-use segment::vector_storage::dense::dynamic_mmap_flags::DynamicMmapFlags;
 use tempfile::tempdir;
 
 const FLAG_COUNT: usize = 50_000_000;
