+++ lib/segment/src/vector_storage/dense/mod.rs
@@ -1,5 +1,4 @@
 pub mod appendable_dense_vector_storage;
-pub mod dynamic_mmap_flags;
 pub mod memmap_dense_vector_storage;
 pub mod mmap_dense_vectors;
 #[cfg(feature = "rocksdb")]
